(APWA-21) [ContactForm, Product, mockuseContactForm, asyncActions, useCategoryCOntent, useCompare, useRemoveProduct, useContactForm, useCookieWarning, useCountries, useFooter, useFooterContact, useCartTrigger, useHeader, useHeaderContact, useMobileMenu, useNavigationTrigger, useWelcomeMessage]
- Refactored talons (Header, Footer, Countries, CookieWarning, Contact, Compare, CategoryContent).
