+++ src/lib/fragments/productItem.graphql
@@ -1,0 +1,84 @@
+fragment ProductItem on ProductInterface {
+    __typename
+    description {
+        html
+    }
+    id
+    media_gallery {
+        label
+        url
+    }
+    name
+    price_range {
+        maximum_price {
+            discount {
+                amount_off
+                percent_off
+            }
+            final_price {
+                currency
+                value
+            }
+            regular_price {
+                currency
+                value
+            }
+        }
+        minimum_price {
+            discount {
+                amount_off
+                percent_off
+            }
+            final_price {
+                currency
+                value
+            }
+            regular_price {
+                currency
+                value
+            }
+        }
+    }
+    sku
+    small_image {
+        url
+        label
+    }
+    url_key
+    url_suffix
+    ... on CustomizableProductInterface {
+        options {
+            __typename
+        }
+    }
+    ... on ConfigurableProduct {
+        configurable_options {
+            attribute_code
+            attribute_id
+            id
+            label
+            values {
+                default_label
+                label
+                store_label
+                use_default_value
+                value_index
+            }
+        }
+        variants {
+            attributes {
+                code
+                value_index
+            }
+            product {
+                id
+                media_gallery {
+                    url
+                    label
+                }
+                sku
+                stock_status
+            }
+        }
+    }
+}
