--- src/peregrine/lib/talons/adeoweb/Compare/useCompare.d.ts
@@ -1,7 +1,0 @@
-import { TProductData } from 'src/peregrine/lib/store/reducers/adeoweb/productCompare';
-
-type TUseCompare = {
-    productData: TProductData;
-};
-
-export function useCompare(): TUseCompare;
