+++ src/peregrine/lib/talons/adeoweb/Compare/useCompare.tsx
@@ -1,6 +1,11 @@
 import { useProductCompareContext } from 'src/peregrine/lib/context/adeoweb/productCompare';
+import { TProductData } from 'src/peregrine/lib/store/reducers/adeoweb/productCompare';
 
-export const useCompare = () => {
+type TUseCompare = {
+    productData: TProductData;
+};
+
+export const useCompare = (): TUseCompare => {
     const [{ productData }] = useProductCompareContext();
 
     return {
