--- src/peregrine/lib/talons/adeoweb/Compare/useRemoveProduct.d.ts
@@ -1,7 +1,0 @@
-import { TProduct } from 'src/lib/types/graphql/Product';
-
-type TUseRemoveProduct = {
-    removeProduct: (product: TProduct) => Promise<void>;
-};
-
-export function useRemoveProduct(): TUseRemoveProduct;
