--- src/peregrine/lib/talons/adeoweb/Compare/useRemoveProduct.js
@@ -1,9 +1,0 @@
-import { useProductCompareContext } from 'src/peregrine/lib/context/adeoweb/productCompare';
-
-export const useRemoveProduct = () => {
-    const [, { removeProduct }] = useProductCompareContext();
-
-    return {
-        removeProduct
-    };
-};
