--- src/peregrine/lib/talons/adeoweb/CookieWarning/useCookieWarning.d.ts
@@ -1,6 +1,0 @@
-export type TUseCookieWarning = {
-    allowCookies: string;
-    handleAllowCookiesClick: () => void;
-};
-
-export function useCookieWarning(): TUseCookieWarning;
