--- src/peregrine/lib/talons/adeoweb/Footer/useFooter.d.ts
@@ -1,5 +1,0 @@
-type TUseFooter = {
-    copyrightText: string;
-};
-
-export function useFooter(): TUseFooter;
