--- src/peregrine/lib/talons/adeoweb/Footer/useFooter.js
@@ -1,9 +1,0 @@
-import { useConfigContext } from '../../../context/adeoweb/config';
-
-export const useFooter = () => {
-    const [{ copyright: copyrightText }] = useConfigContext();
-
-    return {
-        copyrightText: copyrightText || ''
-    };
-};
