+++ src/peregrine/lib/talons/adeoweb/Footer/useFooter.tsx
@@ -1,0 +1,13 @@
+import { useConfigContext } from 'src/peregrine/lib/context/adeoweb/config';
+
+type TUseFooter = {
+    copyrightText: string;
+};
+
+export const useFooter = (): TUseFooter => {
+    const [{ copyright: copyrightText = '' }] = useConfigContext();
+
+    return {
+        copyrightText
+    };
+};
