--- src/peregrine/lib/talons/adeoweb/Footer/useFooterContact.d.ts
@@ -1,8 +1,0 @@
-type TUseuseFooterContact = {
-    storeAddress: string;
-    storePhone: string;
-    storeEmail: string;
-    storeWorkingHours: string;
-};
-
-export function useFooterContact(): TUseuseFooterContact;
