--- src/peregrine/lib/talons/adeoweb/Header/useCartTrigger.d.ts
@@ -1,11 +1,0 @@
-type TUseCartTriggerProps = {
-    createCartMutation: string | DocumentNode;
-    getCartDetailsQuery: string | DocumentNode;
-};
-
-type TUseCartTrigger = {
-    handleClick: () => void;
-    itemCount: number;
-};
-
-export function useCartTrigger(props: TUseCartTriggerProps): TUseCartTrigger;
