--- src/peregrine/lib/talons/adeoweb/Header/useHeader.d.ts
@@ -1,8 +1,0 @@
-type TUseHeader = {
-    handleSearchTriggerClick: () => void;
-    hasBeenOffline: boolean;
-    isOnline: boolean;
-    searchOpen: boolean;
-};
-
-export function useHeader(): TUseHeader;
