+++ src/peregrine/lib/talons/adeoweb/Header/useHeader.tsx
@@ -2,7 +2,14 @@ import { useCallback } from 'react';
 
 import { useAppContext } from 'src/peregrine/lib/context/adeoweb/app';
 
-export const useHeader = () => {
+type TUseHeader = {
+    handleSearchTriggerClick: () => void;
+    hasBeenOffline: boolean;
+    isOnline: boolean;
+    searchOpen: boolean;
+};
+
+export const useHeader = (): TUseHeader => {
     const [{ hasBeenOffline, isOnline, searchOpen }, { toggleSearch }] =
         useAppContext();
