--- src/peregrine/lib/talons/adeoweb/Header/useHeaderContact.d.ts
@@ -1,5 +1,0 @@
-export type TUseHeaderContact = {
-    storePhone: string;
-};
-
-export function useHeaderContact(): TUseHeaderContact;
