+++ src/peregrine/lib/talons/adeoweb/Header/useHeaderContact.tsx
@@ -1,6 +1,10 @@
 import { useConfigContext } from '../../../context/adeoweb/config';
 
-export const useHeaderContact = () => {
+export type TUseHeaderContact = {
+    storePhone: string;
+};
+
+export const useHeaderContact = (): TUseHeaderContact => {
     const [{ store_phone: storePhone }] = useConfigContext();
 
     return { storePhone };
