--- src/peregrine/lib/talons/adeoweb/Header/useMobileMenu.d.ts
@@ -1,6 +1,0 @@
-type TUseMobileMenu = {
-    isOpen: boolean;
-    handleClose: () => void;
-};
-
-export function useMobileMenu(): TUseMobileMenu;
