+++ src/peregrine/lib/talons/adeoweb/Header/useMobileMenu.tsx
@@ -2,7 +2,12 @@ import { useCallback } from 'react';
 
 import { useAppContext } from 'src/peregrine/lib/context/adeoweb/app';
 
-export const useMobileMenu = () => {
+type TUseMobileMenu = {
+    isOpen: boolean;
+    handleClose: () => void;
+};
+
+export const useMobileMenu = (): TUseMobileMenu => {
     const [appState, { closeDrawer }] = useAppContext();
     const { drawer } = appState;
     const isOpen = drawer === 'nav';
