--- src/peregrine/lib/talons/adeoweb/Header/useNavigationTrigger.d.ts
@@ -1,5 +1,0 @@
-type TUseNavigationTrigger = {
-    handleOpenNavigation: () => void;
-};
-
-export function useNavigationTrigger(): TUseNavigationTrigger;
