+++ src/peregrine/lib/talons/adeoweb/Header/useNavigationTrigger.tsx
@@ -2,7 +2,11 @@ import { useCallback } from 'react';
 
 import { useAppContext } from 'src/peregrine/lib/context/adeoweb/app';
 
-export const useNavigationTrigger = () => {
+type TUseNavigationTrigger = {
+    handleOpenNavigation: () => void;
+};
+
+export const useNavigationTrigger = (): TUseNavigationTrigger => {
     const [, { toggleDrawer }] = useAppContext();
 
     const handleOpenNavigation = useCallback(() => {
