--- src/peregrine/lib/talons/adeoweb/Header/useWelcomeMessage.d.ts
@@ -1,5 +1,0 @@
-export type TUseWelcomeMessage = {
-    storeWelcomeMessage: string;
-};
-
-export function useWelcomeMessage(): TUseWelcomeMessage;
