+++ src/peregrine/lib/talons/adeoweb/Header/useWelcomeMessage.tsx
@@ -1,6 +1,10 @@
 import { useConfigContext } from '../../../context/adeoweb/config';
 
-export const useWelcomeMessage = () => {
+export type TUseWelcomeMessage = {
+    storeWelcomeMessage: string;
+};
+
+export const useWelcomeMessage = (): TUseWelcomeMessage => {
     const [{ store_welcome_message: storeWelcomeMessage }] = useConfigContext();
 
     return {
